Refactor import functions into a different file for each one
