+++ src/domain/runner/mod.rs
@@ -5,7 +5,6 @@ pub use self::{
 
 mod abort_data;
 mod bitcoin_network;
-mod call_result;
 mod constants;
 mod contract_runner;
 mod custom_env;
@@ -14,3 +13,4 @@ mod import_functions;
 mod instance_wrapper;
 mod store;
 mod wasmer_runner;
+mod call_result;
